requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

pymupdf==1.23.26

chromadb==0.4.24
sentence-transformers==2.7.0
numpy<2.0

pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# OAuth2 / Keycloak resource server
python-jose[cryptography]==3.3.0
httpx==0.27.0
email-validator==2.1.1

# SQL layer
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Object storage (MinIO / S3-compatible)
minio==7.2.7

# OCR fallback (PaddleOCR). Install paddlepaddle from the PaddlePaddle index:
#   pip install paddlepaddle==2.6.2 -f https://www.paddlepaddle.org.cn/whl/windows/mkl/avx/stable.html
#   pip install paddleocr==2.7.3 Pillow==10.3.0
paddleocr==2.7.3
Pillow==10.3.0
# PaddlePaddle 2.6.x's generated protobuf stubs require protobuf 3.20.x.
# Newer protobuf (4+/5+/6+) triggers "Descriptors cannot be created directly".
# ChromaDB's opentelemetry/googleapis-common-protos only use protobuf for
# telemetry (already disabled in our code path), so this pin is safe.
protobuf==3.20.2

# Auto-grading VLM (Qwen2.5-VL-7B-Instruct). Install PyTorch from the
# CUDA 13.1 wheel index to match the host driver, e.g.:
#   pip install torch>=2.1.0 --index-url https://download.pytorch.org/whl/cu131
torch>=2.1.0
transformers>=4.49.0
accelerate
qwen-vl-utils
bitsandbytes
